Buff raiders base stats and disable raid progression
Doesn't work on servers because it is strictly based on time which
continues to progress regardless if anyone is logged intot the system.
They will still get buffs from scaling difficulty. Tweak as necessary
and if you are playing single player local only and not leaving your
game world running for days at a time feel free to turn it back to true.